Moved bait information into separate inventory script
