+++ Assets/Entity/Player/PlayerInventory.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3ac256712af6da64088bed223b74fc34
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
