+++ Assets/Prefabs/Inventory.prefab
@@ -407,6 +407,7 @@ GameObject:
   - component: {fileID: 2584621981527251822}
   - component: {fileID: 1461224330494846532}
   - component: {fileID: 4945772091463789049}
+  - component: {fileID: 1528095601823452673}
   m_Layer: 5
   m_Name: Inventory
   m_TagString: Untagged
@@ -499,6 +500,24 @@ MonoBehaviour:
   m_BlockingMask:
     serializedVersion: 2
     m_Bits: 4294967295
+--- !u!114 &1528095601823452673
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 3996559547052274909}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 3ac256712af6da64088bed223b74fc34, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  baitTypes: 00000000000000000000000000000000
+  baitText:
+  - {fileID: 7594623609077098205}
+  - {fileID: 9089739003763137432}
+  - {fileID: 8943148669779162416}
+  - {fileID: 8687233094917976268}
 --- !u!1 &5880343684764559122
 GameObject:
   m_ObjectHideFlags: 0
@@ -511,7 +530,7 @@ GameObject:
   - component: {fileID: 5712005830655824647}
   - component: {fileID: 2266086741971821507}
   m_Layer: 5
-  m_Name: Image
+  m_Name: InventoryFrame
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
